+++ src/test/kotlin/com/zepben/auth/client/ZepbenTokenFetcherTest.kt
@@ -22,6 +22,7 @@ import com.zepben.auth.common.StatusCode
 import com.zepben.testutils.auth.TOKEN
 import com.zepben.testutils.exception.ExpectException.expect
 import com.zepben.testutils.vertx.TestHttpServer
+import io.mockk.*
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.core.IsEqual.equalTo
 import org.junit.jupiter.api.AfterEach
@@ -30,7 +31,10 @@ import org.junit.jupiter.api.Test
 import org.mockito.Mockito.*
 import org.mockito.kotlin.mock
 import java.net.http.HttpClient
+import java.net.http.HttpRequest
+import java.net.http.HttpRequest.BodyPublishers
 import java.net.http.HttpResponse
+import javax.net.ssl.SSLContext
 
 internal class ZepbenTokenFetcherTest {
     private lateinit var server: TestHttpServer
@@ -39,23 +43,48 @@ internal class ZepbenTokenFetcherTest {
     private val client = mock<HttpClient>()
     private val response = mock<HttpResponse<String>>()
 
+    private val secureSSLContext = mock<SSLContext>()
+    private val secureConfSSLContext = mock<SSLContext>()
+    private val secureAuthSSLContext = mock<SSLContext>()
+    private val insecureSSLContext = mock<SSLContext>()
+
+    private val secureClient = mock<HttpClient>()
+    private val secureConfClient = mock<HttpClient>()
+    private val secureAuthClient = mock<HttpClient>()
+    private val insecureClient = mock<HttpClient>()
+
+    private val secureTokenFetcher = mock<ZepbenTokenFetcher>()
+    private val insecureTokenFetcher = mock<ZepbenTokenFetcher>()
+
     @BeforeEach
     fun beforeEach() {
         server = TestHttpServer()
         port = server.listen()
         doReturn(response).`when`(client).send(any(), any<HttpResponse.BodyHandler<String>>())
+
+        mockkObject(SSLContextUtils)
+        every { SSLContextUtils.allTrustingSSLContext() } returns insecureSSLContext
+        every { SSLContextUtils.singleCACertSSLContext("confCAFilename") } returns secureConfSSLContext
+        every { SSLContextUtils.singleCACertSSLContext("authCAFilename") } returns secureAuthSSLContext
+
+        mockkStatic(HttpClient::class)
+        every { HttpClient.newHttpClient() } returns secureClient
+        every { HttpClient.newBuilder().sslContext(insecureSSLContext).build() } returns insecureClient
+        every { HttpClient.newBuilder().sslContext(secureConfSSLContext).build() } returns secureConfClient
+        every { HttpClient.newBuilder().sslContext(secureAuthSSLContext).build() } returns secureAuthClient
     }
 
     @AfterEach
     fun afterEach() {
         server.close()
+        unmockkAll()
     }
 
     @Test
     fun testCreateTokenFetcherSuccess() {
         doReturn(StatusCode.OK.code).`when`(response).statusCode()
         doReturn(
-            "{\"authType\": \"AUTH0\", \"audience\": \"test_audience\", \"issuer\": \"test_issuer\"}"
+            "{\"authType\": \"OAUTH\", \"audience\": \"test_audience\", \"issuer\": \"test_issuer\"}"
         ).`when`(response).body()
 
         val tokenFetcher = createTokenFetcher("https://testaddress", confClient = client, authClient = client)
@@ -132,8 +161,7 @@ internal class ZepbenTokenFetcherTest {
         val tokenFetcher = ZepbenTokenFetcher(
             audience = "test_audience",
             issuerDomain = "testissuer.com.au",
-            authMethod = AuthMethod.AUTH0,
-            verifyCertificate = true,
+            authMethod = AuthMethod.OAUTH,
             issuerProtocol = "https",
             tokenPath = "/fake/path",
             client = client
@@ -152,8 +180,7 @@ internal class ZepbenTokenFetcherTest {
         val tokenFetcher = ZepbenTokenFetcher(
             audience = "test_audience",
             issuerDomain = "testissuer.com.au",
-            authMethod = AuthMethod.AUTH0,
-            verifyCertificate = true,
+            authMethod = AuthMethod.OAUTH,
             issuerProtocol = "https",
             tokenPath = "/fake/path",
             client = client
@@ -178,8 +205,7 @@ internal class ZepbenTokenFetcherTest {
         val tokenFetcher = ZepbenTokenFetcher(
             audience = "test_audience",
             issuerDomain = "testissuer.com.au",
-            authMethod = AuthMethod.AUTH0,
-            verifyCertificate = true,
+            authMethod = AuthMethod.OAUTH,
             issuerProtocol = "https",
             tokenPath = "/fake/path",
             client = client
@@ -204,8 +230,7 @@ internal class ZepbenTokenFetcherTest {
         val tokenFetcher = ZepbenTokenFetcher(
             audience = "test_audience",
             issuerDomain = "testissuer.com.au",
-            authMethod = AuthMethod.AUTH0,
-            verifyCertificate = true,
+            authMethod = AuthMethod.OAUTH,
             issuerProtocol = "https",
             tokenPath = "/fake/path",
             client = client
@@ -230,8 +255,7 @@ internal class ZepbenTokenFetcherTest {
         val tokenFetcher = ZepbenTokenFetcher(
             audience = "test_audience",
             issuerDomain = "testissuer.com.au",
-            authMethod = AuthMethod.AUTH0,
-            verifyCertificate = true,
+            authMethod = AuthMethod.OAUTH,
             issuerProtocol = "https",
             tokenPath = "/fake/path",
             client = client
@@ -255,19 +279,107 @@ internal class ZepbenTokenFetcherTest {
             "{\"access_token\":\"$TOKEN\", \"refresh_token\": \"test_refresh_token\", \"token_type\":\"Bearer\"}"
         ).`when`(response).body()
 
-        val tokenFetcher = ZepbenTokenFetcher(
-            audience = "test_audience",
-            issuerDomain = "testissuer.com.au",
-            authMethod = AuthMethod.AUTH0,
-            verifyCertificate = true,
-            issuerProtocol = "https",
-            tokenPath = "/fake/path",
-            client = client,
-            refreshToken = "test_refresh_token"
-        )
-        verify(client, never()).send(any(), any<HttpResponse.BodyHandler<String>>())
-        val token = tokenFetcher.fetchToken()
-        verify(client).send(any(), any<HttpResponse.BodyHandler<String>>())
-        assertThat(token, equalTo("Bearer $TOKEN"))
+        mockStatic(BodyPublishers::class.java, CALLS_REAL_METHODS).use { bodyPublishers ->
+            val tokenFetcher = ZepbenTokenFetcher(
+                audience = "test_audience",
+                issuerDomain = "testissuer.com.au",
+                authMethod = AuthMethod.OAUTH,
+                issuerProtocol = "https",
+                tokenPath = "/fake/path",
+                client = client,
+                refreshToken = "test_refresh_token"
+            )
+            verify(client, never()).send(any(), any<HttpResponse.BodyHandler<String>>())
+            val token = tokenFetcher.fetchToken()
+            verify(client).send(any(), any<HttpResponse.BodyHandler<String>>())
+            bodyPublishers.verify { BodyPublishers.ofString(matches("\"refresh_token\"\\s*:\\s*\"test_refresh_token\"")) }
+            assertThat(token, equalTo("Bearer $TOKEN"))
+        }
+
     }
+
+    @Test
+    fun testConstructorWithVerifyCertificatesOption() {
+        doReturn(response).`when`(secureClient).send(any(), any<HttpResponse.BodyHandler<String>>())
+        doReturn(response).`when`(insecureClient).send(any(), any<HttpResponse.BodyHandler<String>>())
+
+        doReturn(StatusCode.OK.code).`when`(response).statusCode()
+        doReturn("{\"access_token\":\"$TOKEN\", \"token_type\":\"Bearer\"}").`when`(response).body()
+
+        assertThat(
+            ZepbenTokenFetcher("audience", "issuerDomain", AuthMethod.OAUTH, verifyCertificate = true).fetchToken(),
+            equalTo("Bearer $TOKEN")
+        )
+        assertThat(
+            ZepbenTokenFetcher("audience", "issuerDomain", AuthMethod.OAUTH, verifyCertificate = false).fetchToken(),
+            equalTo("Bearer $TOKEN")
+        )
+    }
+
+    @Test
+    fun testConstructorWithCAFilename() {
+        doReturn(response).`when`(secureAuthClient).send(any(), any<HttpResponse.BodyHandler<String>>())
+
+        doReturn(StatusCode.OK.code).`when`(response).statusCode()
+        doReturn("{\"access_token\":\"$TOKEN\", \"token_type\":\"Bearer\"}").`when`(response).body()
+
+        assertThat(
+            ZepbenTokenFetcher("audience", "issuerDomain", AuthMethod.OAUTH, caFilename = "authCAFilename").fetchToken(),
+            equalTo("Bearer $TOKEN")
+        )
+    }
+
+    @Test
+    fun testConstructorWithDefaultTls() {
+        doReturn(response).`when`(secureClient).send(any(), any<HttpResponse.BodyHandler<String>>())
+
+        doReturn(StatusCode.OK.code).`when`(response).statusCode()
+        doReturn("{\"access_token\":\"$TOKEN\", \"token_type\":\"Bearer\"}").`when`(response).body()
+
+        assertThat(
+            ZepbenTokenFetcher("audience", "issuerDomain", AuthMethod.OAUTH).fetchToken(),
+            equalTo("Bearer $TOKEN")
+        )
+    }
+
+    @Test
+    fun testCreateTokenFetcherWithVerifyCertificatesOption() {
+        mockkStatic("com.zepben.auth.client.ZepbenTokenFetcherKt")
+        every {
+            createTokenFetcher("confAddress", secureClient, secureClient, "authTypeField", "audienceField", "issuerDomainField")
+        } returns secureTokenFetcher
+        every {
+            createTokenFetcher("confAddress", insecureClient, insecureClient, "authTypeField", "audienceField", "issuerDomainField")
+        } returns insecureTokenFetcher
+
+        assertThat(createTokenFetcher("confAddress", true, "authTypeField", "audienceField", "issuerDomainField"), equalTo(secureTokenFetcher))
+        assertThat(createTokenFetcher("confAddress", false, "authTypeField", "audienceField", "issuerDomainField"), equalTo(insecureTokenFetcher))
+    }
+
+    @Test
+    fun testCreateTokenFetcherWithCAFilenames() {
+        mockkStatic("com.zepben.auth.client.ZepbenTokenFetcherKt")
+        every {
+            createTokenFetcher("confAddress", secureConfClient, secureAuthClient, "authTypeField", "audienceField", "issuerDomainField")
+        } returns secureTokenFetcher
+
+        assertThat(
+            createTokenFetcher("confAddress", "confCAFilename", "authCAFilename", "authTypeField", "audienceField", "issuerDomainField"),
+            equalTo(secureTokenFetcher)
+        )
+    }
+
+    @Test
+    fun testCreateTokenFetcherWithDefaultTls() {
+        mockkStatic("com.zepben.auth.client.ZepbenTokenFetcherKt")
+        every {
+            createTokenFetcher("confAddress", secureClient, secureClient, "authTypeField", "audienceField", "issuerDomainField")
+        } returns secureTokenFetcher
+
+        assertThat(
+            createTokenFetcher("confAddress", authTypeField="authTypeField", audienceField="audienceField", issuerDomainField="issuerDomainField"),
+            equalTo(secureTokenFetcher)
+        )
+    }
+
 }